fix(sync): use BigInteger for icon_code/color_value, fix snackbar spam
Flutter Color.toARGB32() returns unsigned 32-bit values (up to 4.2B)
which overflow PostgreSQL INTEGER (max 2.1B). Changed to BigInteger
with migration 0006. Also removed persistent sync progress snackbars
that followed across all pages.

+++ backend/app/models/category.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import Column, String, Integer, Float, DateTime, ForeignKey
+from sqlalchemy import Column, String, Integer, BigInteger, Float, DateTime, ForeignKey
 from app.core.db import Base
 
 
@@ -10,8 +10,8 @@ class Category(Base):
     id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
     user_id = Column(String, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     name = Column(String(255), nullable=False)
-    icon_code = Column(Integer, nullable=False)
-    color_value = Column(Integer, nullable=False)
+    icon_code = Column(BigInteger, nullable=False)
+    color_value = Column(BigInteger, nullable=False)
     budget_limit = Column(Float, default=0)
     version = Column(Integer, default=1)
     created_at = Column(
